Implemented model reinstatement and hydration. Tested record saving.
Reinstatement clears the list of changed attributes on a model, declaring its current attributes as unchanged.

Hydration is the same as generation, but all generated models are reinstated.

This has resulted in removing reinstatement from model constructors, so that they are always considered changed even when instantiated. It has also fixed the failing assertions for a couple of association tests.

Records are now also reinstated when successfully saved, and record saving has been unit tested properly.